+++ baccarat/baccarat.go
@@ -45,24 +45,30 @@ func DreamsMenuIntro() (entries map[string][]string) {
 func OnTabSelected(d *dreams.AppObject) {
 	GetBaccTables()
 	BaccRefresh(d)
-	if rpc.Wallet.IsConnected() && bacc.displayed {
-		ActionBuffer(false)
-	}
 }
 
 // Main Baccarat process
 func fetch(d *dreams.AppObject) {
-	bacc.found = true
-	bacc.displayed = true
+	var synced bool
 	time.Sleep(3 * time.Second)
 	for {
 		select {
 		case <-d.Receive():
 			if !rpc.Wallet.IsConnected() || !rpc.Daemon.IsConnected() {
-				disableBaccActions(true)
-				BaccRefresh(d)
+				B.Actions.Hide()
+				synced = false
 				d.WorkDone()
 				continue
+			}
+
+			if !synced {
+				synced = true
+				if FetchLastHand() {
+					B.Front.Objects[0] = showBaccCards()
+					B.Back.Objects[1].(*canvas.Text).Text = bacc.display.result
+					B.Back.Objects[1].Refresh()
+				}
+				B.Actions.Show()
 			}
 
 			fetchBaccSC()
@@ -86,7 +92,6 @@ func ActionBuffer(d bool) {
 		bacc.banker.card1 = 99
 		bacc.banker.card2 = 99
 		bacc.banker.card3 = 99
-		bacc.last = ""
 		bacc.display.result = "Wait for Block..."
 	} else {
 		if rpc.Daemon.IsConnected() && bacc.display.result != "Wait for Block..." {
@@ -97,47 +102,53 @@ func ActionBuffer(d bool) {
 	B.Actions.Refresh()
 }
 
-// Disable Baccarat actions
-func disableBaccActions(d bool) {
-	if d {
-		B.Actions.Hide()
-	} else {
-		B.Actions.Show()
-	}
-
-	B.Actions.Refresh()
-}
-
 // When bet is placed start waiting gif and show txid dialog
-func placeBet(amt, where, title string, w fyne.Window) {
+func placeBet(amt, where, title string, d *dreams.AppObject) {
 	bacc.wait = true
 	ActionBuffer(true)
 	bacc.found = false
-	bacc.displayed = false
+	B.Front.Objects[0] = clearBaccCards()
 	B.Back.Objects[1].(*canvas.Text).Text = bacc.display.result
 	B.Back.Objects[1].Refresh()
-	waiting.Start()
-	waiting.Show()
 	if tx := BaccBet(amt, where); tx == "ID error" {
-		dialog.NewInformation("Baccarat", "Select a table", w).Show()
+		dialog.NewInformation("Baccarat", "Select a table", d.Window).Show()
 	} else {
+		waiting.Start()
+		waiting.Show()
+		go menu.ShowTxDialog(title, "Baccarat", tx, 2*time.Second, d.Window)
 		if tx == "" {
 			bacc.wait = false
 			bacc.found = true
-			bacc.displayed = true
 			bacc.display.result = ""
 			B.Back.Objects[1].(*canvas.Text).Text = bacc.display.result
 			B.Back.Objects[1].Refresh()
-			waiting.Stop()
-			waiting.Hide()
+			stopGif()
+		} else {
+			for i := 0; !bacc.found && i < 100; i++ {
+				FetchHand(tx)
+				time.Sleep(time.Second)
+			}
+
+			stopGif()
+			if bacc.found {
+				B.Front.Objects[0] = showBaccCards()
+				B.Back.Objects[1].(*canvas.Text).Text = bacc.display.result
+				B.Back.Objects[1].Refresh()
+				if !d.IsWindows() {
+					d.Notification("dReams - Baccarat", bacc.display.result)
+				}
+			} else {
+				B.Back.Objects[1].(*canvas.Text).Text = "Error, could not find TX"
+				B.Back.Objects[1].Refresh()
+			}
 		}
-		go menu.ShowTxDialog(title, "Baccarat", tx, 3*time.Second, w)
 	}
+	ActionBuffer(false)
 	bacc.wait = false
 }
 
 // Baccarat action objects
-func baccaratButtons(w fyne.Window) fyne.CanvasObject {
+func baccaratButtons(d *dreams.AppObject) fyne.CanvasObject {
 	entry := dwidget.NewDeroEntry("", 1, 0)
 	entry.PlaceHolder = "dReams:"
 	entry.AllowFloat = false
@@ -170,64 +181,38 @@ func baccaratButtons(w fyne.Window) fyne.CanvasObject {
 	confirmations := widget.NewCheck("Bet confirmation", nil)
 	confirmations.SetChecked(true)
 
-	player_button := widget.NewButton("Player", nil)
-	player_button.Importance = widget.HighImportance
-	player_button.OnTapped = func() {
-		title := fmt.Sprintf("%s %s %s Bet", entry.Text, rpc.GetAssetSCIDName(bacc.assetID), player_button.Text)
+	bet := widget.NewSelect([]string{"Player  (1:1)", "Banker  (95:100)", "Tie  (8:1)"}, nil)
+
+	btnBet := widget.NewButton("Bet", nil)
+	btnBet.Importance = widget.HighImportance
+	btnBet.OnTapped = func() {
+		title := fmt.Sprintf("%s %s %s Bet", entry.Text, rpc.GetAssetSCIDName(bacc.assetID), btnBet.Text)
 		if confirmations.Checked {
 			dialog.NewConfirm("Confirm", title, func(b bool) {
 				if b {
-					placeBet(entry.Text, strings.ToLower(player_button.Text), title, w)
+					go placeBet(entry.Text, strings.ToLower(btnBet.Text), title, d)
 				}
-			}, w).Show()
+			}, d.Window).Show()
 		} else {
-			placeBet(entry.Text, strings.ToLower(player_button.Text), title, w)
+			go placeBet(entry.Text, strings.ToLower(btnBet.Text), title, d)
 		}
 	}
 
-	banker_button := widget.NewButton("Banker", nil)
-	banker_button.Importance = widget.HighImportance
-	banker_button.OnTapped = func() {
-		title := fmt.Sprintf("%s %s %s Bet", entry.Text, rpc.GetAssetSCIDName(bacc.assetID), banker_button.Text)
-		if confirmations.Checked {
-			dialog.NewConfirm("Confirm", title, func(b bool) {
-				if b {
-					placeBet(entry.Text, strings.ToLower(banker_button.Text), title, w)
-				}
-			}, w).Show()
-		} else {
-			placeBet(entry.Text, strings.ToLower(banker_button.Text), title, w)
-		}
+	bet.OnChanged = func(s string) {
+		split := strings.Split(s, "  ")
+		btnBet.Text = split[0]
+		btnBet.Refresh()
 	}
-
-	tie_button := widget.NewButton("Tie", nil)
-	tie_button.Importance = widget.HighImportance
-	tie_button.OnTapped = func() {
-		title := fmt.Sprintf("%s %s %s Bet", entry.Text, rpc.GetAssetSCIDName(bacc.assetID), tie_button.Text)
-		if confirmations.Checked {
-			dialog.NewConfirm("Confirm", title, func(b bool) {
-				if b {
-					placeBet(entry.Text, strings.ToLower(tie_button.Text), title, w)
-				}
-			}, w).Show()
-		} else {
-			placeBet(entry.Text, strings.ToLower(tie_button.Text), title, w)
-		}
-	}
-
-	amt_box := container.NewHScroll(entry)
-	amt_box.SetMinSize(fyne.NewSize(100, 40))
 
 	actions := container.NewVBox(
-		player_button,
-		banker_button,
-		tie_button,
-		amt_box)
+		bet,
+		entry,
+		btnBet)
 
 	var searched string
 	search_entry := widget.NewEntry()
 	search_entry.SetPlaceHolder("TXID:")
-	search_button := widget.NewButton("     Search    ", func() {
+	search_button := widget.NewButtonWithIcon("", dreams.FyneIcon("search"), func() {
 		txid := search_entry.Text
 		if len(txid) == 64 {
 			if txid != searched {
@@ -237,13 +222,16 @@ func baccaratButtons(w fyne.Window) fyne.CanvasObject {
 				B.Back.Objects[1].(*canvas.Text).Text = bacc.display.result
 				B.Back.Objects[1].Refresh()
 				bacc.found = false
-				bacc.displayed = false
-				FetchBaccHand(txid)
+				FetchHand(txid)
+				ActionBuffer(false)
 				if !bacc.found {
 					searched = ""
-					dialog.NewInformation("Search", "No results found", w).Show()
-					ActionBuffer(false)
+					dialog.NewInformation("Search", "No results found", d.Window).Show()
 					bacc.display.result = ""
+					B.Back.Objects[1].(*canvas.Text).Text = bacc.display.result
+					B.Back.Objects[1].Refresh()
+				} else {
+					B.Front.Objects[0] = showBaccCards()
 					B.Back.Objects[1].(*canvas.Text).Text = bacc.display.result
 					B.Back.Objects[1].Refresh()
 				}
@@ -252,7 +240,7 @@ func baccaratButtons(w fyne.Window) fyne.CanvasObject {
 			return
 		}
 
-		dialog.NewInformation("Search", "Not a valid SCID", w).Show()
+		dialog.NewInformation("Search", "Not a valid SCID", d.Window).Show()
 	})
 
 	bacc.display.tableMin = "10"
@@ -272,14 +260,20 @@ func baccaratButtons(w fyne.Window) fyne.CanvasObject {
 
 	search := container.NewVBox(
 		layout.NewSpacer(),
-		container.NewAdaptiveGrid(2,
+		container.NewHBox(
 			container.NewBorder(nil, nil, widget.NewLabel("Table:"), confirmations, table_select),
-			container.NewBorder(nil, nil, nil, search_button, search_entry)))
+			layout.NewSpacer()))
 
 	B.Actions = *container.NewVBox(
 		layout.NewSpacer(),
-		container.NewHBox(layout.NewSpacer(), actions),
-		search)
+		container.NewHBox(
+			search,
+			layout.NewSpacer(),
+			container.NewVBox(
+				layout.NewSpacer(),
+				container.NewBorder(nil, nil, nil, search_button, container.NewStack(dwidget.NewSpacer(560, 0), search_entry))),
+			actions,
+			dwidget.NewSpacer(250, 0)))
 
 	B.Actions.Hide()
 
@@ -290,7 +284,7 @@ func baccaratButtons(w fyne.Window) fyne.CanvasObject {
 func BaccTable(img fyne.Resource) fyne.CanvasObject {
 	table_img := canvas.NewImageFromResource(img)
 	table_img.Resize(fyne.NewSize(1100, 600))
-	table_img.Move(fyne.NewPos(5, 0))
+	table_img.Move(fyne.NewPos(5, 38))
 
 	return container.NewWithoutLayout(table_img)
 }
@@ -319,7 +313,7 @@ func GetBaccTables() {
 			table_names = append(table_names, name)
 		}
 
-		table_select := B.Actions.Objects[2].(*fyne.Container).Objects[1].(*fyne.Container).Objects[0].(*fyne.Container).Objects[0].(*widget.Select)
+		table_select := B.Actions.Objects[1].(*fyne.Container).Objects[0].(*fyne.Container).Objects[1].(*fyne.Container).Objects[0].(*fyne.Container).Objects[0].(*widget.Select)
 		table_select.Options = []string{"dReams"}
 		table_select.Options = append(table_select.Options, table_names...)
 		table_select.Refresh()
@@ -333,13 +327,13 @@ func playerCards(c1, c2, c3 int) fyne.CanvasObject {
 	card3 := holdero.DisplayCard(c3)
 
 	card1.Resize(fyne.NewSize(110, 150))
-	card1.Move(fyne.NewPos(180, 39))
+	card1.Move(fyne.NewPos(180, 94))
 
 	card2.Resize(fyne.NewSize(110, 150))
-	card2.Move(fyne.NewPos(290, 39))
+	card2.Move(fyne.NewPos(290, 94))
 
 	card3.Resize(fyne.NewSize(110, 150))
-	card3.Move(fyne.NewPos(400, 39))
+	card3.Move(fyne.NewPos(400, 94))
 
 	return container.NewWithoutLayout(card1, card2, card3)
 }
@@ -351,13 +345,13 @@ func bankerCards(c1, c2, c3 int) fyne.CanvasObject {
 	card3 := holdero.DisplayCard(c3)
 
 	card1.Resize(fyne.NewSize(110, 150))
-	card1.Move(fyne.NewPos(600, 39))
+	card1.Move(fyne.NewPos(600, 94))
 
 	card2.Resize(fyne.NewSize(110, 150))
-	card2.Move(fyne.NewPos(710, 39))
+	card2.Move(fyne.NewPos(710, 94))
 
 	card3.Resize(fyne.NewSize(110, 150))
-	card3.Move(fyne.NewPos(820, 39))
+	card3.Move(fyne.NewPos(820, 94))
 
 	return container.NewWithoutLayout(card1, card2, card3)
 }
@@ -381,9 +375,6 @@ func showBaccCards() *fyne.Container {
 		playerCards(baccSuit(bacc.player.card1), baccSuit(bacc.player.card2), baccSuit(drawP)),
 		bankerCards(baccSuit(bacc.banker.card1), baccSuit(bacc.banker.card2), baccSuit(drawB)))
 
-	bacc.displayed = true
-	ActionBuffer(false)
-
 	return &content
 }
 
@@ -401,35 +392,11 @@ func BaccRefresh(d *dreams.AppObject) {
 	B.LeftLabel.SetText("Total Hands Played: " + bacc.display.stats.total + "      Player Wins: " + bacc.display.stats.player + "      Ties: " + bacc.display.stats.ties + "      Banker Wins: " + bacc.display.stats.banker + "      Min Bet is " + bacc.display.tableMin + " " + asset_name + ", Max Bet is " + bacc.display.tableMax)
 	B.RightLabel.SetText(asset_name + " Balance: " + rpc.DisplayBalance(asset_name) + "      Dero Balance: " + rpc.DisplayBalance("Dero") + "      Height: " + rpc.Wallet.Display.Height)
 
-	if !bacc.displayed {
-		B.Front.Objects[0] = clearBaccCards()
-		FetchBaccHand(bacc.last)
-		if bacc.found {
-			stopGif()
-			B.Front.Objects[0] = showBaccCards()
-			if !bacc.notified {
-				if !d.IsWindows() {
-					bacc.notified = d.Notification("dReams - Baccarat", bacc.display.result)
-				}
-			}
-		}
-		B.Front.Objects[0].Refresh()
-	}
-
-	if !bacc.wait && rpc.Wallet.Height > bacc.cHeight+4 && !bacc.found && !rpc.IsConfirmingTx() {
-		bacc.display.result = ""
-		ActionBuffer(false)
-		stopGif()
-	}
-
 	if B.Actions.Hidden {
 		if bacc.found && rpc.Wallet.IsConnected() && !rpc.IsConfirmingTx() {
 			ActionBuffer(false)
 		}
 	}
-
-	B.Back.Objects[1].(*canvas.Text).Text = bacc.display.result
-	B.Back.Objects[1].Refresh()
 
 	B.DApp.Refresh()
 }