+++ baccarat/start.go
@@ -28,7 +28,7 @@ import (
 
 const app_tag = "Baccarat"
 
-var version = semver.MustParse("0.3.1")
+var version = semver.MustParse("0.3.1-dev.0")
 var gnomon = gnomes.NewGnomes()
 
 // Check baccarat package version